Split forward mode AD into separate classes
